extract ABI primitives into near-abi-rs
