--- near-sdk-macros/src/core_impl/abi/mod.rs
@@ -1,3 +1,0 @@
-pub mod abi_generator;
-pub mod abi_visitor;
-mod doc;
