+++ near-sdk/src/lib.rs
@@ -70,3 +70,7 @@ pub use serde;
 
 #[doc(hidden)]
 pub use serde_json;
+
+#[doc(hidden)]
+#[cfg(feature = "abi")]
+pub use schemars;
